click to clipboard 2nd commit
