+++ src/components/ClipBoardButton.tsx
@@ -1,0 +1,29 @@
+import React, { useState } from 'react';
+import { FileCopy } from '@mui/icons-material';
+import Tooltip from '@mui/material/Tooltip';
+import IconButton from '@mui/material/IconButton';
+
+const ClipboardButton = ({ textToCopy }) => {
+  const [isCopied, setIsCopied] = useState(false);
+  const [timeoutId, setTimeoutId] = useState(null);
+
+  const handleCopyToClipboard = () => {
+    clearTimeout(timeoutId); // Clear the previous timeout
+    navigator.clipboard.writeText(textToCopy);
+    setIsCopied(true);
+    const newTimeoutId = setTimeout(() => {
+      setIsCopied(false);
+    }, 2000);
+    setTimeoutId(newTimeoutId); // Store the new timeout ID
+  };
+
+  return (
+    <Tooltip title={isCopied ? 'Copied' : 'Copy to clipboard'} placement="bottom">
+      <IconButton onClick={handleCopyToClipboard}>
+        <FileCopy style={{ fontSize: '16px' }} />
+      </IconButton>
+    </Tooltip>
+  );
+};
+
+export default ClipboardButton;
